+++ crates/sevenmark_formatter/src/format/brace/code.rs
@@ -1,13 +1,17 @@
 use pretty::{Arena, DocAllocator, DocBuilder};
 use sevenmark_ast::CodeElement;
 
+use crate::FormatConfig;
 use crate::format::params::format_params;
 
-pub fn format_code<'a>(a: &'a Arena<'a>, e: &CodeElement) -> DocBuilder<'a, Arena<'a>> {
+pub fn format_code<'a>(
+    a: &'a Arena<'a>,
+    e: &CodeElement,
+    config: &FormatConfig,
+) -> DocBuilder<'a, Arena<'a>> {
     a.text("{{{#code")
-        .append(format_params(a, &e.parameters))
+        .append(format_params(a, &e.parameters, config))
         .append(a.hardline())
-        .append(a.text(e.value.trim().to_string()))
-        .append(a.hardline())
+        .append(a.text(e.value.clone()))
         .append(a.text("}}}"))
 }