+++ crates/sevenmark_formatter/src/format/brace/literal.rs
@@ -1,10 +1,15 @@
 use pretty::{Arena, DocAllocator, DocBuilder};
 use sevenmark_ast::LiteralElement;
 
+use crate::FormatConfig;
 use crate::format::element::format_elements;
 
-pub fn format_literal<'a>(a: &'a Arena<'a>, e: &LiteralElement) -> DocBuilder<'a, Arena<'a>> {
+pub fn format_literal<'a>(
+    a: &'a Arena<'a>,
+    e: &LiteralElement,
+    config: &FormatConfig,
+) -> DocBuilder<'a, Arena<'a>> {
     a.text("{{{")
-        .append(format_elements(a, &e.children))
+        .append(format_elements(a, &e.children, config))
         .append(a.text("}}}"))
 }